+++ packages/orm/angel_migration_runner/lib/src/mariadb/runner.dart
@@ -7,6 +7,7 @@ import '../runner.dart';
 import '../util.dart';
 import 'schema.dart';
 
+/// A MariaDB database migration runner.
 class MariaDbMigrationRunner implements MigrationRunner {
   final _log = Logger('MariaDbMigrationRunner');